+++ CHANGELOG.md
@@ -10,6 +10,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ---
 
+## [1.3.X] - 2020-12-XX
+
+### Changed
+- Small changes in `index.html` in example
+- Remove dependency `astar-typescript` in example
+- Add new controller in example for diagonal movement
+
+---
+
 ## [1.2.5] - 2020-12-13
 
 ### Fixed